accountService/requirements.txt
django==4.2.1
djangorestframework==3.14.0
psycopg2-binary==2.9.6
pyjwt==2.6.0
requests==2.28.2
pika==1.3.1